requirements.txt
paho-mqtt
wyzeapy
python-dotenv